refactor(knowledge-store): make the adapter the package's service.py
The store's consumer-facing verbs sat in app/services as a loose file
named after the mechanism, while the rest of the subsystem lived in
app/knowledge_store. Callers had to know both places, and the next verb
would have been guessed into whichever the author saw first.

Now it reads like file_storage: the package owns its application service,
and consumers import one module. Also says so in the docstring, since the
reason a caller reaches past a facade is usually that nobody told them it
was one.
